+++ app/Core/SiteAssetException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Copot\Core;
+
+class SiteAssetException extends \RuntimeException
+{
+}